+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+## 2.5.0
+* gui.authorize supports now async, see [API reference](https://github.com/ma-ha/easy-web-app/blob/master/API-Reference.md) 
+* dynamicHeader/dynamicFooter/dynamicTitle/dynamicRow : added pageName as parameter
+
 ## 2.4.2
 * table/list: text more robust, can also consume numbers